env_requirements.txt
# ============================================================================
# WAKEE RELOADED - Requirements (Anaconda wr_env Python 3.11)
# ============================================================================

# HuggingFace
huggingface-hub==0.20.3

# ML/DL
torch==2.1.2
torchvision==0.16.2
onnx==1.15.0
onnxruntime==1.16.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Cloud
boto3==1.34.34

# Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Streamlit
streamlit==1.31.0

# CV
opencv-python==4.9.0.80
Pillow==10.2.0

# Data
pandas==2.2.0
numpy==1.26.3

# MLOps
evidently==0.4.14
mlflow==2.10.2

# LLM
# langchain==0.1.4
# langchain-mistralai==0.0.4

# Utils
python-dotenv==1.0.1